requirements.txt
altgraph==0.17
future==0.18.2
img2pdf==0.4.0
lxml==4.6.2
pefile==2019.4.18
pikepdf==2.8.0
Pillow==8.1.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyPDF2==1.26.0
pywin32-ctypes==0.2.0
